+++ src/argilla_llama_index/llama_index_handler.py
@@ -5,11 +5,12 @@ from packaging.version import parse
 from typing import Any, Dict, List, Optional
 import warnings
 
+from argilla_llama_index.helpers import _calc_time, _get_time_diff
+
 import argilla as rg
 from argilla._constants import DEFAULT_API_KEY, DEFAULT_API_URL
 from llama_index.core.callbacks.base_handler import BaseCallbackHandler
 from llama_index.core.callbacks.schema import (
-    BASE_TRACE_EVENT,
     CBEventType,
     EventPayload,
     CBEvent,
@@ -669,40 +670,3 @@ class ArgillaCallbackHandler(BaseCallbackHandler):
 
         event = CBEvent(event_type, payload=payload, id_=event_id)
         self.events_data[event_id].append(event)
-
-
-# Auxiliary methods
-def _get_time_diff(event_1_time_str: str, event_2_time_str: str) -> float:
-    """
-    Get the time difference between two events Follows the American format (month, day, year).
-
-    Args:
-        event_1_time_str (str): The first event time.
-        event_2_time_str (str): The second event time.
-
-    Returns:
-        float: The time difference between the two events.
-    """
-    time_format = "%m/%d/%Y, %H:%M:%S.%f"
-
-    event_1_time = datetime.strptime(event_1_time_str, time_format)
-    event_2_time = datetime.strptime(event_2_time_str, time_format)
-
-    return round((event_2_time - event_1_time).total_seconds(), 4)
-
-
-def _calc_time(events_data: Dict[str, List[CBEvent]], id: str) -> float:
-    """
-    Calculate the time difference between the start and end of an event using the events_data.
-
-    Args:
-        events_data (Dict[str, List[CBEvent]]): The events data, stored in a dictionary.
-        id (str): The event id to calculate the time difference between start and finish timestamps.
-
-    Returns:
-        float: The time difference between the start and end of the event.
-    """
-
-    start_time = events_data[id][0].time
-    end_time = events_data[id][1].time
-    return _get_time_diff(start_time, end_time)
